// src/utils/libraryPaths.ts
import { LibraryTab } from 'types/libraryTab';
import * as userSettings from 'scripts/settings/userSettings';

interface LibraryViewDefinition {
    index: number
    label: string
    view: LibraryTab
    isDefault?: boolean
}

export interface LibraryRoute {
    path: string,
    views: LibraryViewDefinition[]
}

export const LibraryRoutes: LibraryRoute[] = [
    {
        path: '/livetv',
        views: [
            {
                index: 0,
                label: 'Programs',
                view: LibraryTab.Programs,
                isDefault: true
            },
            {
                index: 1,
                label: 'Guide',
                view: LibraryTab.Guide
            },
            {
                index: 2,
                label: 'Channels',
                view: LibraryTab.Channels
            },
            {
                index: 3,
                label: 'Recordings',
                view: LibraryTab.Recordings
            },
            {
                index: 4,
                label: 'Schedule',
                view: LibraryTab.Schedule
            },
            {
                index: 5,
                label: 'Series',
                view: LibraryTab.SeriesTimers
            }
        ]
    },
    {
        path: '/movies',
        views: [
            {
                index: 0,
                label: 'Movies',
                view: LibraryTab.Movies,
                isDefault: true
            },
            {
                index: 1,
                label: 'Suggestions',
                view: LibraryTab.Suggestions
            },
            {
                index: 2,
                label: 'Favorites',
                view: LibraryTab.Favorites
            },
            {
                index: 3,
                label: 'Collections',
                view: LibraryTab.Collections
            },
            {
                index: 4,
                label: 'Genres',
                view: LibraryTab.Genres
            }
        ]
    },
    {
        path: '/music',
        views: [
            {
                index: 0,
                label: 'Albums',
                view: LibraryTab.Albums,
                isDefault: true
            },
            {
                index: 1,
                label: 'Suggestions',
                view: LibraryTab.Suggestions
            },
            {
                index: 2,
                label: 'HeaderAlbumArtists',
                view: LibraryTab.AlbumArtists
            },
            {
                index: 3,
                label: 'Artists',
                view: LibraryTab.Artists
            },
            {
                index: 4,
                label: 'Playlists',
                view: LibraryTab.Playlists
            },
            {
                index: 5,
                label: 'Songs',
                view: LibraryTab.Songs
            },
            {
                index: 6,
                label: 'Genres',
                view: LibraryTab.Genres
            }
        ]
    },
    {
        path: '/tv',
        views: [
            {
                index: 0,
                label: 'Shows',
                view: LibraryTab.Series,
                isDefault: true
            },
            {
                index: 1,
                label: 'Suggestions',
                view: LibraryTab.Suggestions
            },
            {
                index: 2,
                label: 'TabUpcoming',
                view: LibraryTab.Upcoming
            },
            {
                index: 3,
                label: 'Genres',
                view: LibraryTab.Genres
            },
            {
                index: 4,
                label: 'TabNetworks',
                view: LibraryTab.Networks
            },
            {
                index: 5,
                label: 'Episodes',
                view: LibraryTab.Episodes
            }
        ]
    },
    {
        path: '/homevideos',
        views: [
            {
                index: 0,
                label: 'Photos',
                view: LibraryTab.Photos,
                isDefault: true
            },
            {
                index: 1,
                label: 'HeaderPhotoAlbums',
                view: LibraryTab.PhotoAlbums,
                isDefault: true
            },
            {
                index: 2,
                label: 'HeaderVideos',
                view: LibraryTab.Videos
            }
        ]
    },
    {
        path: '/books',
        views: [
            {
                index: 0,
                label: 'Books',
                view: LibraryTab.Books,
                isDefault: true
            }
        ]
    }
];

/**
 * Utility function to check if a path is a library path.
 */
export const isLibraryPath = (path: string) => (
    LibraryRoutes.some(route => route.path === path)
);

/**
 * Utility function to get the default view index for a specified URL path and library.
 */
export const getDefaultViewIndex = (path: string, libraryId?: string | null) => {
    if (!libraryId) return 0;

    const views = LibraryRoutes.find(route => route.path === path)?.views ?? [];
    const defaultView = userSettings.get('landing-' + libraryId, false);

    return views.find(view => view.view === defaultView)?.index
        ?? views.find(view => view.isDefault)?.index
        ?? 0;
};
